修复图像检测不精确的 bug
